update model createupdatedelete
# Conflicts:
#	cmd/server.go
#	internal/handler/user.go
#	internal/repo/user.go
#	internal/service/coingecko.go
